requirements.txt
openai 
langchain-pinecone
langchain-openai 
python-dotenv